requirements.txt
dotenv==0.9.9
mss==10.0.0
pillow==11.3.0
python-dotenv==1.1.1
